# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "API Documentation for User Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get all notifications for the authenticated user",
                "operationId": "0f031f66cd8c52bf73e89fb2d050ad83",
                "responses": {
                    "200": {
                        "description": "List of notifications"
                    },
                    "404": {
                        "description": "No notifications found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a new notification for the authenticated user",
                "operationId": "258ba7c348fd4283c6b3802a69fe2fb0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message",
                                    "type"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "New message received"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "info"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get a specific notification by ID",
                "operationId": "0c8e017ab6d765f82f97368b80f52b8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Update a notification",
                "operationId": "38c5f13acd3d0b84545d309539e6b972",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Updated notification message"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "warning"
                                    },
                                    "is_read": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete a notification",
                "operationId": "ccdcc730b2088097dd94128f24960dbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/quiz-questions": {
            "get": {
                "tags": [
                    "Quiz Questions"
                ],
                "summary": "Get all quiz questions",
                "operationId": "022690e965fd7da8e3ff36c81ab7249e",
                "responses": {
                    "200": {
                        "description": "List of quiz questions"
                    },
                    "404": {
                        "description": "No quiz questions found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Quiz Questions"
                ],
                "summary": "Create a new quiz question",
                "operationId": "c7aa8b3bc0be8d57e383bbed081039a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question",
                                    "answer",
                                    "quiz_id"
                                ],
                                "properties": {
                                    "question": {
                                        "type": "string",
                                        "example": "What is the capital of France?"
                                    },
                                    "answer": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "quiz_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quiz question created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/quiz-questions/{id}": {
            "get": {
                "tags": [
                    "Quiz Questions"
                ],
                "summary": "Get a specific quiz question",
                "operationId": "39c78e6bb3cc5ae0eab6fcfc3985e122",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz question ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz question data"
                    },
                    "404": {
                        "description": "Quiz question not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Quiz Questions"
                ],
                "summary": "Update a quiz question",
                "operationId": "b97be638015f79b3a01addc532ab1a99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz question ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "question": {
                                        "type": "string",
                                        "example": "Updated question?"
                                    },
                                    "answer": {
                                        "type": "string",
                                        "example": "Updated answer"
                                    },
                                    "quiz_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz question updated successfully"
                    },
                    "404": {
                        "description": "Quiz question not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Quiz Questions"
                ],
                "summary": "Delete a quiz question",
                "operationId": "b52e4db0e43ccb93b4861113fd8b977d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz question ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz question deleted successfully"
                    },
                    "404": {
                        "description": "Quiz question not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}